fix: routing stability, map centering, and trip persistence
- Fixed TomTom routing error (MAP_MATCHING_FAILURE: Origin 0,0) by adding coordinates to Backend DTOs and Frontend models.
- Implemented resilient geocoding fallback via Nominatim API for legacy trips.
- Resolved "0 rows affected" concurrency exception in partitioned trips using ExecuteUpdateAsync and timestamp precision (HasPrecision(6)).
- Fixed React hook violation in TripDetailsModal.
- Improved map positioning with dynamic centering (ChangeView) and intelligent AutoBounds.

+++ Backend/src/SmartLogist.Application/DTOs/Trip/UpdateTripDto.cs
@@ -8,7 +8,7 @@ public class UpdateTripDto
     public DateTime? ActualDeparture { get; set; }
     public DateTime? ActualArrival { get; set; }
     public string? Notes { get; set; }
-    public double? ActualFuelConsumption { get; set; }
+    public float? ActualFuelConsumption { get; set; }
     public int? Rating { get; set; }
     public string? ManagerReview { get; set; }
 }
