+++ Backend/src/SmartLogist.Application/DTOs/Vehicle/VehicleDtos.cs
@@ -9,11 +9,11 @@ public class VehicleDto
     public string LicensePlate { get; set; } = string.Empty;
     public string Type { get; set; } = string.Empty;
     public string FuelType { get; set; } = string.Empty;
-    public double FuelConsumption { get; set; }
-    public double Height { get; set; }
-    public double Width { get; set; }
-    public double Length { get; set; }
-    public double Weight { get; set; }
+    public float FuelConsumption { get; set; }
+    public float Height { get; set; }
+    public float Width { get; set; }
+    public float Length { get; set; }
+    public float Weight { get; set; }
     public bool IsHazardous { get; set; }
     public string Status { get; set; } = string.Empty;
     public DateTime CreatedAt { get; set; }
@@ -21,9 +21,9 @@ public class VehicleDto
     public int? AssignedDriverId { get; set; }
     public bool HasPendingDeletion { get; set; }
     public bool HasPendingUpdate { get; set; }
-    public double TotalMileage { get; set; }
-    public double MileageAtLastMaintenance { get; set; }
-    public double KmUntilMaintenance { get; set; }
+    public float TotalMileage { get; set; }
+    public float MileageAtLastMaintenance { get; set; }
+    public float KmUntilMaintenance { get; set; }
 }
 
 public class CreateVehicleDto
@@ -32,15 +32,15 @@ public class CreateVehicleDto
     public string LicensePlate { get; set; } = string.Empty;
     public string Type { get; set; } = string.Empty;
     public string FuelType { get; set; } = string.Empty;
-    public double FuelConsumption { get; set; }
-    public double Height { get; set; }
-    public double Width { get; set; }
-    public double Length { get; set; }
-    public double Weight { get; set; }
+    public float FuelConsumption { get; set; }
+    public float Height { get; set; }
+    public float Width { get; set; }
+    public float Length { get; set; }
+    public float Weight { get; set; }
     public bool IsHazardous { get; set; }
     public VehicleStatus Status { get; set; } = VehicleStatus.Available;
-    public double TotalMileage { get; set; } = 0;
-    public double MileageAtLastMaintenance { get; set; } = 0;
+    public float TotalMileage { get; set; } = 0;
+    public float MileageAtLastMaintenance { get; set; } = 0;
 }
 
 public class UpdateVehicleDto
@@ -49,15 +49,15 @@ public class UpdateVehicleDto
     public string LicensePlate { get; set; } = string.Empty;
     public string Type { get; set; } = string.Empty;
     public string FuelType { get; set; } = string.Empty;
-    public double FuelConsumption { get; set; }
-    public double Height { get; set; }
-    public double Width { get; set; }
-    public double Length { get; set; }
-    public double Weight { get; set; }
+    public float FuelConsumption { get; set; }
+    public float Height { get; set; }
+    public float Width { get; set; }
+    public float Length { get; set; }
+    public float Weight { get; set; }
     public bool IsHazardous { get; set; }
     public VehicleStatus Status { get; set; }
-    public double TotalMileage { get; set; }
-    public double MileageAtLastMaintenance { get; set; }
+    public float TotalMileage { get; set; }
+    public float MileageAtLastMaintenance { get; set; }
 }
 
 public class AssignVehicleDto
